% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectoRdb.R
\name{cos.sim}
\alias{cos.sim}
\title{cos.sim
Find the cosine similarity as defined in http://en.wikipedia.org/wiki/Cosine_similarity}
\usage{
\method{cos}{sim}(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
the cosine similarity between the two vectors
}
\description{
cos.sim
Find the cosine similarity as defined in http://en.wikipedia.org/wiki/Cosine_similarity
}
