% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectoRdb.R
\name{insert}
\alias{insert}
\alias{insert.VectorDB}
\title{insert
Insert a vector into a VectorDB object}
\usage{
insert(x, ...)

\method{insert}{VectorDB}(x, k, v)
}
\arguments{
\item{x}{VectorDB object}

\item{...}{other arguments}

\item{k}{key (name)}

\item{v}{vector}
}
\value{
the VectorDB object `x` with the new element `k-->v` concatenated in the end
}
\description{
insert
Insert a vector into a VectorDB object
}
\examples{
x <- VectorDB()
for(i in letters[1:10]) { x <- insert(x, i, sample(1:4, 5, replace=TRUE)) }
}
