% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectoRdb.R
\name{retrieve}
\alias{retrieve}
\title{retrieve
Retrieve an entry from a VectorDB object using the key}
\usage{
retrieve(x, ...)
}
\arguments{
\item{x}{VectorDB object}

\item{...}{other arguments}
}
\description{
retrieve
Retrieve an entry from a VectorDB object using the key
}
