% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectoRdb.R
\name{VectorDB}
\alias{VectorDB}
\title{VectorDB
S3 class VectorDB constructor}
\usage{
VectorDB()
}
\value{
an empty VectorDB object
}
\description{
VectorDB
S3 class VectorDB constructor
}
\examples{
x <- VectorDB()
}
